Add top level module for padding
